chore: migrate to typescript plugin
