--- packages/language-server/bin/doodl-language-server.js
@@ -1,8 +1,0 @@
-#!/usr/bin/env node
-if (process.argv.includes("--version")) {
-	const pkgJSON = require("../package.json");
-	console.log(`${pkgJSON["version"]}`);
-}
-else {
-	require("../out/index.js");
-}
