--- packages/language-server/package.json
@@ -1,22 +1,0 @@
-{
-	"name": "doodl-language-server",
-	"version": "0.0.1",
-	"main": "out/index.js",
-	"files": [
-		"out/**/*.js",
-		"out/**/*.d.ts"
-	],
-	"bin": {
-		"doodl-language-server": "./bin/doodl-language-server.js"
-	},
-	"dependencies": {
-		"@volar/language-core": "~2.4.23",
-		"@volar/language-server": "~2.4.23",
-		"@volar/language-service": "~2.4.23",
-		"@volar/typescript": "~2.4.23",
-		"volar-service-html": "volar-2.4",
-		"volar-service-typescript": "volar-2.4",
-		"vscode-html-languageservice": "^5.5.1",
-		"vscode-uri": "^3.1.0"
-	}
-}
