--- packages/language-server/tsconfig.json
@@ -1,10 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"outDir": "out",
-		"rootDir": "src"
-	},
-	"include": [
-		"src"
-	],
-}
