+++ packages/language-service/src/index.ts
@@ -1,0 +1,2 @@
+export { createDoodlLanguagePlugin, DoodlVirtualCode } from './language-plugin.js';
+export { createDoodlServicePlugin } from './service-plugin.js';
