+++ packages/language-service/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "include": ["src"],
+
+  "compilerOptions": {
+    "declaration": true,
+    "outDir": "lib"
+  }
+}
