+++ packages/typescript-plugin/lib/index.cjs
@@ -1,0 +1,18 @@
+'use strict'
+
+const {
+  createDoodlLanguagePlugin,
+  createDoodlServicePlugin
+} = require('doodl-language-service')
+const {
+  createLanguageServicePlugin
+} = require('@volar/typescript/lib/quickstart/createLanguageServicePlugin.js')
+
+const plugin = createLanguageServicePlugin((ts, info) => {
+  return {
+    languagePlugins: [createDoodlLanguagePlugin()],
+    servicePlugins: [createDoodlServicePlugin()]
+  }
+})
+
+module.exports = plugin
