+++ packages/vscode/package.json
@@ -57,6 +57,16 @@
 					"source.graphql": "graphql"
 				}
 			}
+		],
+		"typescriptServerPlugins": [
+			{
+				"name": "doodl-typescript-plugin",
+				"enableForWorkspaceTypeScriptVersions": true,
+				"configNamespace": "typescript",
+				"languages": [
+					"doodl"
+				]
+			}
 		]
 	},
 	"scripts": {
@@ -65,13 +75,12 @@
 		"pack": "npm run build && vsce package --pre-release",
 		"release": "npm run build && vsce publish --pre-release"
 	},
+	"dependencies": {
+		"doodl-typescript-plugin": "workspace:*"
+	},
 	"devDependencies": {
 		"@types/vscode": "^1.104.0",
-		"@volar/language-server": "~2.4.23",
-		"@volar/vscode": "~2.4.23",
-		"doodl-language-server": "workspace:*",
 		"esbuild": "^0.25.10",
-		"vsce": "^2.15.0",
-		"vscode-languageclient": "^9.0.1"
+		"vsce": "^2.15.0"
 	}
 }
