+++ packages/vscode/tsconfig.json
@@ -6,10 +6,5 @@
 	},
 	"include": [
 		"src",
-	],
-	"references": [
-		{
-			"path": "../language-server/tsconfig.json"
-		}
 	]
 }
