+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "target": "ES2020",
+    "module": "ESNext",
+    "moduleResolution": "node",
+    "strict": true,
+    "plugins": [
+      { "name": "doodl-typescript-plugin" }
+    ]
+  },
+  "include": ["*.dood", "sample/*.dood"]
+}
